+++ application/views/Part/homepage.php
@@ -1,0 +1,5 @@
+<div class="row">
+    {parts}
+    <div class="span4"><a href="/part/gimme/{partID}"><img src="/data/parts/{pic}" /></a></div>
+    {/parts}
+</div>