+++ application/views/Part/justonepart.php
@@ -1,0 +1,10 @@
+<div class="row">
+    <div class="span3"><img src="/data/parts/{pic}" title="{partCode}"/></div>
+    <div class="span8 offset1">
+        <p class="lead">Part Code: {partCode}</p><br/>
+        <p class="text-left">Built Date: {dateTime}</p>
+        <p class="text-left">Built Plant: {plantID}</p>
+        <p class="text-left">Price: {cost}</p>
+        
+    </div>
+</div>