+++ application/views/Robot/homepage.php
@@ -1,0 +1,5 @@
+<div class="row">
+    {robots}
+    <div class="span4"><a href="/data/bots/{location}"><img src="/data/bots/{location}" /></a></div>
+    {/robots}
+</div>