+++ application/views/_robotCell.php
@@ -1,0 +1,5 @@
+
+<a href="/data/bots/{location}">
+    <img src="/data/bots/{location}"  />
+</a>
+<input type="checkbox" name="partCheckBox" value="{id}" class="checkbox">