+++ ChangeLog.md
@@ -11,4 +11,11 @@ change log (newest last)
 
 * added bots and parts in data directory by Lei
 
-* 
+* finished robot controller and model by Lei
+
+* finished TransacPartsRecords and PurchasePartsRecords model by Jing
+
+* finished parts controller and view by Jing
+
+* add parts model, assembly view and controller by Rachel
+